Discover the platform packages that write rows
Three were named by hand — settings, extsecrets, keyvault — and the list
was short by two. platform/events writes event_outbox and platform/jobs
deletes river_job, so neither write was ever compared against the
ownership map, and event_outbox IS declared there. A hand-kept list of
writers is one somebody has to remember to extend, and nothing fails
when they do not: the same hole vault_secret sat in until somebody
noticed.

The set is derived now — every package under internal/platform holding a
row-writing statement — and pinned, because a discovery that stopped
finding writers would report a clean tree in the same words as a tree
whose writers are all declared. Widening to the whole of
internal/platform would be the other mistake: it would judge packages
that write nothing and have no declaration to judge them against.

storekit is excluded, and that is not an exemption: its writes name
whatever table the CALLER passes, so counting it would attribute every
module's versioned write to one package — and each of those is already
judged where the table is named.

The two writes the derivation found are answered rather than assumed.
river_job is owned by the fleet that runs it, and its history is
elsewhere: what a job DID is audited by the module that did it, and the
purge that deletes these rows is audited by the erasure that ordered it.
The relay's outbox mark is ratified where it happens — storekit writes
the row and the relay retires it, which is what an outbox IS, and
marking published_at in the transaction that shipped is what makes a
crash re-ship rather than lose.

The discovery lives in its own file: the ownership gate asks whether a
package writes only what it owns, and this asks who the writers are.
